refactor: rename @types folders to types
