+++ __tests__/cryptoWallet/utils/request.test.ts
@@ -1,8 +1,8 @@
 /* eslint-disable import/first */
 import 'jest'
 
-import { BLOCKCHAIN_NAMESPACES } from 'features/blockchain/constants'
-import { isSupportedCaipNamespace } from 'features/caip/utils/isSupportedCaipNamespace'
+import { BLOCKCHAIN_NAMESPACES } from '~/features/blockchain/constants'
+import { isSupportedCaipNamespace } from '~/features/caip/utils/isSupportedCaipNamespace'
 
 // Required as request.ts imports a simple function from features/caip but the whole internal module is loaded even though it's unnecessary for these unit tests
 jest.mock('features/caip', () => ({
