+++ src/features/blockchain/eip155/hooks/useBlockchainRequestHandlers.Eip155.ts
@@ -1,18 +1,18 @@
 import { Web3WalletTypes } from '@walletconnect/web3wallet'
 import axios from 'axios'
 import { ethers } from 'ethers'
-import { useChainMetadatasCustom } from 'features/blockchain/hooks'
-import { ChainMetadatas } from 'features/caip'
 import * as React from 'react'
 import { fromZodError } from 'zod-validation-error'
 
-import { SupportedBlockchainNamespace } from '~/features/blockchain/types/enums'
+import { ChainMetadatas } from '~/features/caip'
 
+import { useChainMetadatasCustom } from '../../hooks'
+import { SupportedBlockchainNamespace } from '../../types'
 import {
   AddEthereumChainRequestParams,
   BlockchainRequestHandlersEip155,
   Eip155RpcMethod,
-} from '../@types'
+} from '../types'
 
 // HACK: Prevent invalid requests from being generated by legacy frontends.
 const adjustTransactionData = async ({
