+++ src/features/blockchain/eip155/index.ts
@@ -1,3 +1,3 @@
-export * from './@types'
+export * from './types'
 export * from './utils'
 export * from './hooks'
