+++ src/features/blockchain/near/index.ts
@@ -1,4 +1,4 @@
-export * from './@types'
+export * from './types'
 export * from './constants'
 export * from './hooks'
 export * from './utils'
