+++ src/features/blockchain/near/utils/nearCreateViewAccessKey.ts
@@ -1,7 +1,7 @@
 import { providers } from 'near-api-js'
 import { AccessKeyView } from 'near-api-js/lib/providers/provider'
 
-import { NearAccountPointer } from '../@types'
+import { NearAccountPointer } from '../types'
 
 export function nearCreateViewAccessKey({
   provider,
