+++ src/features/blockchain/near/utils/nearSignTransactions.ts
@@ -1,6 +1,6 @@
 import { InMemorySigner, transactions } from 'near-api-js'
 
-import { NearAccount } from '../@types'
+import { NearAccount } from '../types'
 
 export const nearSignTransactions = async ({
   nearAccount,
