+++ src/features/walletConnect/utils/mockAddEthereumChainRequest.ts
@@ -3,7 +3,8 @@ import { ChainId } from 'caip'
 import {
   AddEthereumChainRequestParam,
   Eip155RpcMethod,
-} from '../../blockchain/eip155/@types'
+} from '~/features/blockchain/eip155'
+
 import { WalletConnectRequestParams } from '../types'
 
 export function mockAddEthereumChainRequest({
